Put the Either adapter in separate package. Add KDocs
