+++ arrow-integrations-retrofit-adapter/src/test/kotlin/arrow/integrations/retrofit/adapter/retrofit/SuspedApiClientTest.kt
@@ -1,7 +1,7 @@
 package arrow.integrations.retrofit.adapter.retrofit
 
 import arrow.core.Either
-import arrow.integrations.retrofit.adapter.ResponseE
+import arrow.integrations.retrofit.adapter.either.ResponseE
 import arrow.integrations.retrofit.adapter.mock.ErrorMock
 import arrow.integrations.retrofit.adapter.mock.ResponseMock
 import retrofit2.http.GET
